# swagger.yaml
openapi: 3.0.0
info:
  title: DeepThought_NudgeAPI by Manish Kumar Panda
  contact: {}
  version: '1.0'
servers:
# - url: http://localhost:5000/api/v3/app
- url: https://deepthought-nudge-api.onrender.com/
  variables: {}
paths:
...
  /images/{id}/upload-image:
    put:
      tags:
      - Events
      summary: Update Event Image
      description: 64783a7e3e6c3f6a30160330
      operationId: UpdateEventImage
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      requestBody:
        content:
          multipart/form-data:
            encoding: {}
            schema:
              required:
              - image
              type: object
              properties:
                image:
                  type: string
                  format: binary
        required: false
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /images/upload-image:
    post:
      tags:
      - Events
      summary: Add Event Image
      operationId: AddEventImage
      parameters: []
      requestBody:
        content:
          multipart/form-data:
            encoding: {}
            schema:
              required:
              - image
              type: object
              properties:
                image:
                  type: string
                  format: binary
        required: false
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /users:
    post:
      tags:
      - User
      summary: Create User
      operationId: CreateUser
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/CreateUserRequest'
              - example:
                  name: ''
                  role: attendee
            example:
              name: ''
              role: attendee
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    get:
      tags:
      - User
      summary: Get All Users
      operationId: GetAllUsers
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
  /users/{id}:
    put:
      tags:
      - User
      summary: Update User
      operationId: UpdateUser
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/UpdateUserRequest'
              - example:
                  name: ''
                  role: attendee
            example:
              name: ''
              role: attendee
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    delete:
      tags:
      - User
      summary: Delete User
      operationId: DeleteUser
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
  /subCategory:
    post:
      tags:
      - SubCategory
      summary: Create SubCategory
      operationId: CreateSubCategory
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/CreateSubCategoryRequest'
              - example:
                  name: ''
                  parentCategory: ''
            example:
              name: ''
              parentCategory: ''
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    get:
      tags:
      - SubCategory
      summary: Get All SubCategories
      operationId: GetAllSubCategories
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
  /subCategory/{id}:
    put:
      tags:
      - SubCategory
      summary: Update SubCategory
      operationId: UpdateSubCategory
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/UpdateSubCategoryRequest'
              - example:
                  name: ''
                  parentCategory: ''
            example:
              name: ''
              parentCategory: ''
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    delete:
      tags:
      - SubCategory
      summary: Delete SubCategory
      operationId: DeleteSubCategory
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    get:
      tags:
      - SubCategory
      summary: Get Single Subcategory
      operationId: GetSingleSubcategory
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
  /nudge:
    post:
      tags:
      - Nudge
      summary: Create Nudge
      operationId: CreateNudge
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/CreateNudgeRequest'
              - example:
                  eventCategoryId: ''
                  eventId: ''
                  title: ''
                  icon: ''
                  scheduled:
                  - on: 2025-08-11
                    from: 2025-08-11 9:00:00
                    to: 2025-08-11 13:00:00
                  coverImage: ''
                  invitationLine: ''
            example:
              eventCategoryId: ''
              eventId: ''
              title: ''
              icon: ''
              scheduled:
              - on: 2025-08-11
                from: 2025-08-11 9:00:00
                to: 2025-08-11 13:00:00
              coverImage: ''
              invitationLine: ''
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    get:
      tags:
      - Nudge
      summary: Get All Nudges
      operationId: GetAllNudges
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
  /nudge/{id}:
    put:
      tags:
      - Nudge
      summary: Update Nudge
      description: 'try this id : 6477b922a6dce41d2b0a0c31'
      operationId: UpdateNudge
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/UpdateNudgeRequest'
              - example:
                  eventCategoryId: ''
                  eventId: ''
                  title: ''
                  icon: ''
                  scheduledOn: ''
                  scheduledFrom: ''
                  scheduledTo: ''
                  coverImage: ''
                  invitationLine: ''
            example:
              eventCategoryId: ''
              eventId: ''
              title: ''
              icon: ''
              scheduledOn: ''
              scheduledFrom: ''
              scheduledTo: ''
              coverImage: ''
              invitationLine: ''
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    delete:
      tags:
      - Nudge
      summary: Delete Nudge
      description: deleting any nudge will remove the nudge from database & you might need to change default values in other endpoints in order to search for existing nudges
      operationId: DeleteNudge
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    get:
      tags:
      - Nudge
      summary: Get Single Nudge By Id
      operationId: GetSingleNudgeById
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
  /nudge/upload-image:
    post:
      tags:
      - Nudge
      summary: Upload Nudge Cover Image
      operationId: UploadNudgeCoverImage
      parameters: []
      requestBody:
        content:
          multipart/form-data:
            encoding: {}
            schema:
              required:
              - image
              type: object
              properties:
                image:
                  type: string
                  format: binary
        required: false
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /nudge/image/{id}/upload-image:
    put:
      tags:
      - Nudge
      summary: Update Uploaded Nudge Image Cover
      operationId: UpdateUploadedNudgeImageCover
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      requestBody:
        content:
          multipart/form-data:
            encoding: {}
            schema:
              required:
              - image
              type: object
              properties:
                image:
                  type: string
                  format: binary
        required: false
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
components:
  schemas:
...
    Scheduled:
      title: Scheduled
      required:
      - on
      - from
      - to
      type: object
      properties:
        on:
          type: string
        from:
          type: string
        to:
          type: string
      example:
        on: 2023-10-08
        from: 2023-10-08 9:00:00
        to: 2023-10-08 13:00:00
...
    CreateUserRequest:
      title: CreateUserRequest
      required:
      - name
      - role
      type: object
      properties:
        name:
          type: string
        role:
          type: string
      example:
        name: ''
        role: attendee
    UpdateUserRequest:
      title: UpdateUserRequest
      required:
      - name
      - role
      type: object
      properties:
        name:
          type: string
        role:
          type: string
      example:
        name: ''
        role: attendee
    CreateSubCategoryRequest:
      title: CreateSubCategoryRequest
      required:
      - name
      - parentCategory
      type: object
      properties:
        name:
          type: string
        parentCategory:
          type: string
      example:
        name: ''
        parentCategory: ''
    UpdateSubCategoryRequest:
      title: UpdateSubCategoryRequest
      required:
      - name
      - parentCategory
      type: object
      properties:
        name:
          type: string
        parentCategory:
          type: string
      example:
        name: ''
        parentCategory: ''
    CreateNudgeRequest:
      title: CreateNudgeRequest
      required:
      - eventCategoryId
      - eventId
      - title
      - icon
      - scheduled
      - coverImage
      - invitationLine
      type: object
      properties:
        eventCategoryId:
          type: string
        eventId:
          type: string
        title:
          type: string
        icon:
          type: string
        scheduled:
          type: array
          items:
            $ref: '#/components/schemas/Scheduled'
          description: ''
        coverImage:
          type: string
        invitationLine:
          type: string
      example:
        eventCategoryId: ''
        eventId: ''
        title: ''
        icon: ''
        scheduled:
        - on: 2025-08-11
          from: 2025-08-11 9:00:00
          to: 2025-08-11 13:00:00
        coverImage: ''
        invitationLine: ''
    UpdateNudgeRequest:
      title: UpdateNudgeRequest
      required:
      - eventCategoryId
      - eventId
      - title
      - icon
      - scheduledOn
      - scheduledFrom
      - scheduledTo
      - coverImage
      - invitationLine
      type: object
      properties:
        eventCategoryId:
          type: string
        eventId:
          type: string
        title:
          type: string
        icon:
          type: string
        scheduledOn:
          type: string
        scheduledFrom:
          type: string
        scheduledTo:
          type: string
        coverImage:
          type: string
        invitationLine:
          type: string
      example:
        eventCategoryId: ''
        eventId: ''
        title: ''
        icon: ''
        scheduledOn: ''
        scheduledFrom: ''
        scheduledTo: ''
        coverImage: ''
        invitationLine: ''
...
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
- httpBearer: []
...